add customer abstraction
